+++ tests/testthat/test-xpose_data.R
@@ -92,5 +92,9 @@ test_that('properly handles errors in files', {
 
 test_that('properly uses mapped ID column', {
   xpdb_1 <- xpose_data(file = 'run004.lst', dir = 'data', quiet = TRUE)
+  expect_equal(parse_nm_input_record(xpdb_1$code), c(ID="FOO"),
+               info="The input record wants to remap FOO to ID")
+  expect_true("ID" %in% names(xpdb_1$data$data[[1]]),
+              info="FOO is remapped to ID accurately")
 })