+++ NEWS.md
@@ -2,7 +2,8 @@
 ### General
 * Address CRAN requests
 * Improved internal structure of VPC
-* Fixed bug with vpc 1.0.1
+* Fixed bug with vpc 1.0.1'
+* Allow for remapped column names in model file (like ID=SUBJID)
 
 # xpose 0.4.1
 ### General